+++ lib/src/watch_box_builder.dart
@@ -1,80 +1,40 @@
 part of hive_flutter;
 
 /// Signature for a function that builds a widget given a [Box].
-typedef BoxWidgetBuilder = Widget Function(BuildContext context, Box box);
+typedef BoxWidgetBuilder<T> = Widget Function(BuildContext context, Box<T> box);
 
 /// A general-purpose widget which rebuilds itself when the box or a specific
 /// key change.
-class WatchBoxBuilder extends StatefulWidget {
+class WatchBoxBuilder<T> extends StatelessWidget {
   /// Creates a widget that rebuilds itself when a value in the [box] changes.
   ///
-  /// If you specify [watchKeys], the widget only refreshes when a value
-  /// associated to a key in [watchKeys] changes.
+  /// If you specify [watchKey], the widget only refreshes when a value
+  /// associated to a key in [watchKey] changes.
   WatchBoxBuilder({
     Key key,
     @required this.box,
     @required this.builder,
-    this.watchKeys,
+    this.watchKey,
   })  : assert(box != null),
         assert(builder != null),
         super(key: key);
 
   /// The box which should be watched.
-  final Box box;
+  final Box<T> box;
 
   /// Called every time the box changes. The builder must not return null.
   final BoxWidgetBuilder builder;
 
   /// Specifies which keys should be watched.
-  final List<String> watchKeys;
-
-  @override
-  _WatchBoxBuilderState createState() => _WatchBoxBuilderState();
-}
-
-class _WatchBoxBuilderState extends State<WatchBoxBuilder> {
-  @visibleForTesting
-  StreamSubscription subscription;
-
-  @override
-  void initState() {
-    super.initState();
-
-    _subscribe();
-  }
-
-  @override
-  void didUpdateWidget(WatchBoxBuilder oldWidget) {
-    super.didUpdateWidget(oldWidget);
-
-    if (widget.box != oldWidget.box) {
-      _unsubscribe();
-      _subscribe();
-    }
-  }
-
-  void _subscribe() {
-    subscription = widget.box.watch().listen((event) {
-      if (widget.watchKeys != null && !widget.watchKeys.contains(event.key)) {
-        return;
-      }
-
-      setState(() {});
-    });
-  }
-
-  void _unsubscribe() {
-    subscription?.cancel();
-  }
+  final String watchKey;
 
   @override
   Widget build(BuildContext context) {
-    return widget.builder(context, widget.box);
-  }
-
-  @override
-  void dispose() {
-    _unsubscribe();
-    super.dispose();
+    return ValueListenableBuilder(
+      valueListenable: box.listenable(key: watchKey),
+      builder: (BuildContext context, Box<T> value, Widget child) {
+        return builder(context, box);
+      },
+    );
   }
 }